RUSSIAN ROULETTE(C)
Completely re-did the program after refreshing my knowledge with some professional help (YouTube) and then began to try to set the singular bullet into place so the game remembers to only register one. 

This was because I kept running into a weird bug where spinning the barrel led to firing for some reason. It is fixed now though so no need to go on about that. 

This is a rudimentary Russian Roulette game. I am POSITIVE there is room for improvement but this is what I came up with for my personal challenge.